pyproject.toml
[project]
name = "yunabot"
version = "0.1.0"
description = ""
authors = [
    {name = "cool-pants",email = "muban.sg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "d20 (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
